+++ xpra/gtk/configure/home.py
@@ -46,7 +46,6 @@ class HomeGUI(BaseGUIWindow):
         self.sub("Features", "features.png", "Enable or disable feature groups", "features")
         self.sub("Settings", "gears.png", "Configure specific settings", "settings")
         self.sub("Picture compression", "encoding.png", "Encodings, speed and quality", "encodings", "xpra.codecs")
-        # self.sub("GStreamer", "gstreamer.png", "Configure the GStreamer codecs", "gstreamer")
         self.sub("Debugging", "bugs.png", "Configure debug logging", "debug")
 
         self.sub("Shadow Server", "shadow.png", "Configure the Shadow Server", "shadow", "xpra.server")